Move angle helpers to path utils
